scraper/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
entsoe-py==0.3.2
idna==2.10
numpy==1.20.1
pandas==1.2.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.23
tabulate==0.8.9
urllib3==1.26.3
